+++ src/main/java/com/example/proyectospring/Repositories/ProductRepository.java
@@ -4,4 +4,4 @@ package com.example.proyectospring.Repositories;
 import com.example.proyectospring.Models.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {}
+public interface ProductRepository extends CrudRepository<Product, Long> {}
